Base Library Code
Branching out code from another project. This is the first broken out version of this code and is therefore still in need of quite a bit of improvement. We need closer mimicing of the base Sequelize object, which hopefully we'll get over time.

The overall goal of this library is to have a fully test compatible version of the Sequelize API that can be used as a drop in for mocking and unit testing.