Fix `Util.stack` functionality
Stack looks like it is a getter which is where `prepareStackTrace` is actually called. Order matters for getting that and resetting the `prepareStackTrace` function.

TODO: Test has been converted to a bit more of a integration test than unit test to make sure we get coverage of this. Unit test code remains in place but commented out and if/when we get an integration or end-to-end test suite, we should transfer this test over to one of those suites and uncomment the unit test code here.